requirements.txt
openai
langchain
chromadb
streamlit
pypdf
tiktoken
python-dotenv
